requirements.txt
astroid==1.4.7
coverage==3.7.1
dateutils==0.6.6
Django==1.7.8
django-cors-headers==0.13
django-extensions==1.4.5
django-filter==0.8
djangorestframework==3.0.5
ipython==2.3.0
lazy-object-proxy==1.2.2
logilab-common==1.2.2
pep8==1.5.7
psycopg2==2.5.4
pycparser==2.10
pyflakes==0.8.1
pylint==1.4.1
python-dateutil==2.5.3
pytz==2014.10
requests==2.5.0
six==1.8.0
Werkzeug==0.9.6
wheel==0.24.0
wrapt==1.10.8
